+++ music_assistant/providers/snapcast/provider.py
@@ -1,0 +1,60 @@
+"""Snapcast Player Provider implementation."""
+
+from __future__ import annotations
+
+import asyncio
+from enum import StrEnum
+from typing import TYPE_CHECKING
+
+from music_assistant.models.player_provider import PlayerProvider
+
+if TYPE_CHECKING:
+    from snapcast.control.server import Snapserver
+
+    from .player import SnapcastPlayer
+
+
+class SnapCastStreamType(StrEnum):
+    """Snapcast stream types."""
+
+    MUSIC = "music"
+    ANNOUNCEMENT = "announcement"
+
+
+class SnapcastPlayerProvider(PlayerProvider):
+    """Snapcast Player Provider for synchronized audio playback."""
+
+    def __init__(self, *args, **kwargs) -> None:
+        """Initialize the provider."""
+        super().__init__(*args, **kwargs)
+        self.snapcast: Snapserver | None = None
+        self._players: dict[str, SnapcastPlayer] = {}
+        self._stream_tasks: dict[str, asyncio.Task] = {}
+
+    async def handle_async_init(self) -> None:
+        """Handle async initialization of the provider."""
+        # Initialize snapcast server connection
+        # This would involve setting up the snapcast server and client connections
+
+    async def unload(self, is_removed: bool = False) -> None:
+        """Handle unload/close of the provider."""
+        if self.snapcast:
+            # Clean up snapcast connections
+            pass
+
+    def _get_stream_name(self, player_id: str, stream_type: SnapCastStreamType) -> str:
+        """Get stream name for a player and stream type."""
+        return f"ma_{player_id}_{stream_type.value}"
+
+    async def _get_or_create_stream(self, stream_name: str, queue_id: str):
+        """Get or create a snapcast stream."""
+        # Implementation would create/get snapcast streams
+
+    def _get_snapgroup(self, player_id: str):
+        """Get snapcast group for a player."""
+        # Implementation would return the snapcast group
+
+    async def poll_player(self, player_id: str) -> None:
+        """Poll player for state updates."""
+        if player := self._players.get(player_id):
+            await player.poll()
